+++ CHANGELOG.md
@@ -9,9 +9,11 @@ Not Released Yet
 ## Popup
 - New side panel that is more interactive and gives a refreshed look to the app (can be activated by clicking the hamburger menu, or switched back to the old interface)
 - New About page with information regarding the App, and more
+- Added scripts uptime status to the extension popup homepage
 
 ## Bug Fix
 - Fixed issue where, on Youtube, highlighted content wouldn't update when clicking on a new video
+- Fixed issue where some kanji would be highlighted as learned, when it was not the case
 
 # Changelog v0.5.6
 Not Released Yet